move to doubles for higher value precision

+++ HReader/HRBMI.m
@@ -8,25 +8,26 @@
 
 #import "HRBMI.h"
 
-
 @implementation HRBMI
 
 - (BOOL)isNormal {
-    float bmi = self.value;
+    double bmi = self.value;
     return !(bmi < [self normalLow] || bmi > [self normalHigh]);
-
 }
+
 - (NSString *)leftValue {
-    
     return [NSString stringWithFormat:@"%0.1f", self.value];
 }
+
 - (NSString *)rightValue {
     return [NSString stringWithFormat:@"%0.1f-%0.1f", self.normalLow, self.normalHigh];
 }
-- (float)normalLow {
+
+- (double)normalLow {
     return 18.5;
 }
-- (float)normalHigh {
+
+- (double)normalHigh {
     return 22.9;
 }
